+++ app/region_codes.py
@@ -1,5 +1,3 @@
-# region_codes.py
-
 SIDO_CODE_MAP = {
     "11": "서울",
     "26": "부산",
@@ -10,14 +8,14 @@ SIDO_CODE_MAP = {
     "31": "울산",
     "36": "세종",
     "41": "경기",
-    "51": "강원특별자치도",
+    "51": "강원",
     "43": "충청북도",
     "44": "충청남도",
     "45": "전라북도",
     "46": "전라남도",
     "47": "경상북도",
     "48": "경상남도",
-    "50": "제주특별자치도"
+    "50": "제주"
 }
 
 GUGUN_CODE_MAP = {
@@ -135,25 +133,25 @@ GUGUN_CODE_MAP = {
     "4180": "경기 연천군",
     "4182": "경기 가평군",
     "4183": "경기 양평군",
-    "5100": "강원특별자치도",
-    "5111": "강원특별자치도 춘천시",
-    "5113": "강원특별자치도 원주시",
-    "5115": "강원특별자치도 강릉시",
-    "5117": "강원특별자치도 동해시",
-    "5119": "강원특별자치도 태백시",
-    "5121": "강원특별자치도 속초시",
-    "5123": "강원특별자치도 삼척시",
-    "5172": "강원특별자치도 홍천군",
-    "5173": "강원특별자치도 횡성군",
-    "5175": "강원특별자치도 영월군",
-    "5176": "강원특별자치도 평창군",
-    "5177": "강원특별자치도 정선군",
-    "5178": "강원특별자치도 철원군",
-    "5179": "강원특별자치도 화천군",
-    "5180": "강원특별자치도 양구군",
-    "5181": "강원특별자치도 인제군",
-    "5182": "강원특별자치도 고성군",
-    "5183": "강원특별자치도 양양군",
+    "5100": "강원",
+    "5111": "강원 춘천시",
+    "5113": "강원 원주시",
+    "5115": "강원 강릉시",
+    "5117": "강원 동해시",
+    "5119": "강원 태백시",
+    "5121": "강원 속초시",
+    "5123": "강원 삼척시",
+    "5172": "강원 홍천군",
+    "5173": "강원 횡성군",
+    "5175": "강원 영월군",
+    "5176": "강원 평창군",
+    "5177": "강원 정선군",
+    "5178": "강원 철원군",
+    "5179": "강원 화천군",
+    "5180": "강원 양구군",
+    "5181": "강원 인제군",
+    "5182": "강원 고성군",
+    "5183": "강원 양양군",
     "4300": "충청북도",
     "4311": "충청북도 청주시",
     "4313": "충청북도 충주시",
@@ -263,9 +261,9 @@ GUGUN_CODE_MAP = {
     "4887": "경상남도 함양군",
     "4888": "경상남도 거창군",
     "4889": "경상남도 합천군",
-    "5000": "제주특별자치도",
-    "5011": "제주특별자치도 제주시",
-    "5013": "제주특별자치도 서귀포시"
+    "5000": "제주",
+    "5011": "제주 제주시",
+    "5013": "제주 서귀포시"
 }
 
 def get_region_name(signgucode: str, signgucodesub: str = None) -> str:
